scalability/requirements.txt
ansible>=2.10.7
cbor>=1.0.0
GitPython>=3.1.24
pybars3>=0.9.7
python-gflags>=3.1.2
PyYAML>=5.4.1
requests>=2.26.0
six>=1.16.0
termcolor>=1.1.0
absl-py>=0.14.1
